Move price level enum, fix spacing
